use portable-atomic for AtomicU64 where needed
AtomicU64 isn't available in libstd on all targets. portable-atomic
fills the gap, and can be nicely cfg-guarded to only be used on such
targets.

without this patch, gix-status (and thus a lot of other gix crates, and
cargo) cannot be built on Debian armel.

technically, AtomicUsize is also not guaranteed to exist for all
targets, and a similar patch could be written to support those as well
if desired.

Signed-off-by: Fabian Grünbichler <git@fabian.gruenbichler.email>