Simplify path construction and config; improve comments
This stops attempting to make the Windows-specific helper function
semi-portable. I had done that with the idea tha it could be tested
on all platforms so as to gain better coverage in practice. But it
doesn't really make sense to do this because it keeps the native
path separator `\` from being used even where it is now clearer to
use it, and because such tests would not be able to use common
values such as `C:\Program Files` because a path that starts with
`C:\`, while absolute on Windows, is relative on Unix-like systems.