Scope env vars to the job definition that needs them
The variables that controlled the behavior of Rust programs and the
output of `cargo` are only used in the `build-release` job, not in
the `create-release` jobs, and one of them was already set in the
more suitable narrower scope. This moves them to that scope,
without duplication. They still apply to all steps in the build
jobs.

This also:

- Adjusts comment wording to clarify how `CARGO` is set and used.

- Quotes `1` as an environment variable value, since such values
  are always strings (it was being converted), but having it as a
  numeric YAML literal made it seem like that was not the case.

- Adds spacing for clarity and for stylistic consistency across the
  two job definitions.